Modifying spawn() syscall, pipes now working between processes
This commit changes the 'flag' parameter of spawn() to a pointer to
an array of 3 ints, which let the caller specify file descriptors to
start the process with. Each call down from vfs_spawn() also has
an added 'fds' parameter, so that init_task can copy the needed
process objects.

This was done so that pipes (or other file descriptors, of course)
could be used as stdin, stdout and stderr for the new process.

(also now trying to be more descriptive in commit messages :P)